Modified to get MonkeyMan working on scala 2.9.2 and sbt 0.11.3
Removed the Proguard plugin; don't know how to update it, and don't
need it right now.

Also added project/Build.scala to define a get-jars task; this
gets a list of all the dependency jars used by a project, and is
used to make the launcher script.

Also added the eclipse plugin to create and update .project and
.classpath files for Eclipse.